+++ src/style/themes/mint/mint-theme.config.js
@@ -21,6 +21,7 @@ export default (palette) => {
       return {
         colorsActionMajor500: this.colors.primary,
         colorsActionMajor600: this.colors.secondary,
+        colorsActionMajor150: this.colors.loadingBarBackground,
       };
     },
   };